pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "crewai>=0.177.0",
    "crewai-tools>=0.69.0",
    "flake8>=7.3.0",
    "google-genai>=1.33.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.1.10",
    "markdown>=3.9",
    "openai>=1.106.1",
    "pydantic>=2.11.7",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "streamlit>=1.49.1",
]
